+++ frontend/libs/studio-pure-functions/src/ArrayUtils/ArrayUtils.test.ts
@@ -1,4 +1,4 @@
-import { ArrayUtils } from './ArrayUtilss';
+import { ArrayUtils } from './ArrayUtils';
 
 describe('ArrayUtils', () => {
   describe('removeDuplicates', () => {
