+++ frontend/libs/studio-pure-functions/src/ArrayUtils/index.ts
@@ -1,0 +1,1 @@
+export * from './ArrayUtils';
