--- frontend/libs/studio-pure-functions/src/ArrayUtilss/index.ts
@@ -1,1 +1,0 @@
-export * from './ArrayUtilss';
